@@ -2,28 +2,26 @@
 	"name": "sin",
 	"version": "0.0.0",
 	"private": true,
+	"type": "module",
 	"scripts": {
-		"pretty": "prettier --write ./**/*.ts",
-		"test": "jest --no-cache --verbose --runInBand ./lib/",
-		"test-watch": "jest --no-cache --verbose --watch-all --runInBand ./lib/",
-		"cov": "jest --no-cache --verbose --runInBand --coverage --collect-coverage-from './lib/**/*.ts' ./lib/",
-		"cov-watch": "jest --no-cache --verbose --watch-all --runInBand --coverage --collect-coverage-from './lib/**/*.ts' ./lib/"
+		"pretty": "prettier --write ./**/*.ts ./**/*.json ./**/*.mjs",
+		"typecheck": "tsc --noEmit",
+		"cov-report": "tap report html --allow-incomplete-coverage",
+		"test": "npm run pretty && npm run typecheck && tap --jobs 4 --allow-incomplete-coverage && npm run cov-report"
 	},
 	"// UPGRADE_POINT": {},
 	"dependencies": {
-		"@types/jest": "27.0.3",
-		"@types/better-sqlite3": "7.4.1",
-		"@types/glob": "7.2.0",
-		"better-sqlite3": "7.4.5",
-		"chalk": "4.1.2",
-		"commander": "8.3.0",
-		"jest": "27.4.3",
-		"ts-jest": "27.1.1",
+		"@types/better-sqlite3": "7.6.8",
+		"@types/node": "20.10.7",
+		"better-sqlite3": "9.2.2",
+		"chalk": "5.3.0",
+		"commander": "11.1.0",
 		"js-logger": "1.6.1",
-		"prettier": "2.5.1",
-		"prettier-plugin-organize-imports": "2.3.4",
-		"ts-node": "10.4.0",
-		"typescript": "4.5.2",
-		"zx": "4.2.0"
+		"prettier": "3.1.1",
+		"prettier-plugin-organize-imports": "3.2.4",
+		"tap": "18.6.1",
+		"tsx": "4.7.0",
+		"typescript": "5.3.3",
+		"zx": "7.2.3"
 	}
 }
