@@ -2,7 +2,8 @@ import type { Database as DatabaseT, Options, Statement } from 'better-sqlite3'
 import Database from 'better-sqlite3'
 import chalk from 'chalk'
 import { existsSync } from 'fs'
-import type { AnalysedFile, AnalysedFileRow, BasicLogger } from './types'
+
+import type { AnalysedFile, AnalysedFileRow, BasicLogger } from './types.js'
 
 interface Config {
 	sqlitePath: string
@@ -14,8 +15,7 @@ interface Config {
 export class LocalDatabase {
 	sqlite: DatabaseT
 	logger: BasicLogger
-	allowedSpecificLicenses: string[] = []
-	allowedLicenseCategories: string[] = []
+	allowedLicenses: string[] = []
 	stmtInsertFile: Statement
 	stmtUpdateFile: Statement
 
@@ -24,7 +24,9 @@ export class LocalDatabase {
 
 		if (!existsSync(config.sqlitePath)) {
 			this.logger.info(
-				chalk`{magenta Database ${config.sqlitePath} does not exist, creating it.}`,
+				chalk.magenta(
+					`Database ${config.sqlitePath} does not exist, creating it.`,
+				),
 			)
 		}
 		const opts: Options = {}
@@ -49,6 +51,7 @@ export class LocalDatabase {
 				:file_path,
 				:content_sha256,
 				:content_text,
+				:scancode_entry,
 				:licenses,
 				:previous_accepted_reason,
 				:current_accepted_reason,
@@ -63,6 +66,7 @@ export class LocalDatabase {
 				content_sha256 = :content_sha256,
 				content_text = :content_text,
 				licenses = :licenses,
+				scancode_entry = :scancode_entry,
 				previous_accepted_reason = current_accepted_reason,
 				current_accepted_reason = NULL,
 				current_accepted_at = NULL,
@@ -92,6 +96,7 @@ export class LocalDatabase {
 				file_path TEXT PRIMARY KEY,
 				content_sha256 TEXT,
 				content_text TEXT,
+				scancode_entry TEXT,
 				licenses TEXT,
 				previous_accepted_reason TEXT,
 				current_accepted_reason TEXT,
@@ -101,14 +106,7 @@ export class LocalDatabase {
 		`)
 
 		this.sqlite.exec(`
-			CREATE TABLE IF NOT EXISTS accepted_license_categories
-			(
-				name TEXT PRIMARY KEY
-			)
-		`)
-
-		this.sqlite.exec(`
-			CREATE TABLE IF NOT EXISTS accepted_license_names
+			CREATE TABLE IF NOT EXISTS allowed_licenses
 			(
 				name TEXT PRIMARY KEY
 			)
@@ -118,12 +116,10 @@ export class LocalDatabase {
 	}
 
 	loadGlobalSettings() {
-		this.allowedSpecificLicenses = this.fetchAllowedSpecificLicenses()
-		this.allowedLicenseCategories = this.fetchAllowedLicenseCategories()
+		this.allowedLicenses = this.fetchAllowedLicenses()
 	}
 
 	fetchAllAnalysedFiles() {
-		// NOTE: Skipping content_text, because it's big and irrelevant.
 		const stmt = this.sqlite.prepare(`
 			SELECT
 				file_path,
@@ -138,17 +134,19 @@ export class LocalDatabase {
 		return this.fetchAnalysedFiles(stmt)
 	}
 
-	fetchAnalysedFilesNeedingInvestigation() {
-		// NOTE: Skipping content_text, because it's big and irrelevant.
+	fetchAnalysedFilesNeedingInvestigation(verbose: boolean) {
+		const columns = [
+			'file_path',
+			// 'content_sha256',
+			'licenses',
+			'previous_accepted_reason',
+			'current_accepted_reason',
+			'current_accepted_at',
+			'is_legal_document',
+			...(verbose ? ['scancode_entry'] : []),
+		].join(', ')
 		const stmt = this.sqlite.prepare(`
-			SELECT
-				file_path,
-				content_sha256,
-				licenses,
-				previous_accepted_reason,
-				current_accepted_reason,
-				current_accepted_at,
-				is_legal_document
+			SELECT ${columns}
 			FROM analysed_files
 			WHERE licenses IS NOT NULL
 			AND current_accepted_reason IS NULL
@@ -192,6 +190,9 @@ export class LocalDatabase {
 		if (row.previous_accepted_reason) {
 			file.previousAcceptedReason = row.previous_accepted_reason
 		}
+		if (row.scancode_entry) {
+			file.scanCodeEntry = JSON.parse(row.scancode_entry)
+		}
 		return file
 	}
 
@@ -200,11 +201,15 @@ export class LocalDatabase {
 		const licenses = file.licenses?.length
 			? JSON.stringify(file.licenses)
 			: null
+		const scanCodeEntry = file.scanCodeEntry
+			? JSON.stringify(file.scanCodeEntry)
+			: null
 		if (exists) {
 			this.stmtUpdateFile.run({
 				file_path: file.filePath,
 				content_sha256: file.contentSha256,
 				content_text: file.contentText,
+				scancode_entry: scanCodeEntry,
 				licenses,
 				is_legal_document: isLegalDocument,
 			})
@@ -213,6 +218,7 @@ export class LocalDatabase {
 				file_path: file.filePath,
 				content_sha256: file.contentSha256,
 				content_text: file.contentText,
+				scancode_entry: scanCodeEntry,
 				licenses,
 				previous_accepted_reason: null,
 				current_accepted_reason: null,
@@ -222,25 +228,10 @@ export class LocalDatabase {
 		}
 	}
 
-	fetchAllowedSpecificLicenses() {
+	fetchAllowedLicenses() {
 		const stmt = this.sqlite.prepare(`
 			SELECT name
-			FROM accepted_license_names
-		`)
-		const rows = stmt.all() as {
-			name: string
-		}[]
-		const names = []
-		for (const row of rows) {
-			names.push(row.name)
-		}
-		return names
-	}
-
-	fetchAllowedLicenseCategories() {
-		const stmt = this.sqlite.prepare(`
-			SELECT name
-			FROM accepted_license_categories
+			FROM allowed_licenses
 		`)
 		const rows = stmt.all() as {
 			name: string
