@@ -1,7 +1,8 @@
 import chalk from 'chalk'
 import { existsSync } from 'fs'
 import { $ } from 'zx'
-import type { BasicLogger } from './types'
+
+import type { BasicLogger } from './types.js'
 
 interface Config {
 	dirtyRoot: string
@@ -25,20 +26,32 @@ export class ScanStep2 {
 
 	async run() {
 		this.logger.info(
-			chalk`{yellow === STEP 2: Extract any archives among dirty files}`,
+			chalk.yellow(`=== STEP 2: Extract any archives among dirty files`),
 		)
+
+		const extractcodeBinaryPath = 'extractcode'
 
 		if (!existsSync(this.dirtyRoot)) {
 			this.logger.info(`Nothing to be done (no dirty files).`)
 			return
 		}
 		if (this.skipExtractArchives) {
-			this.logger.info(chalk`{red SKIPPING extracting archives.}`)
+			this.logger.info(
+				chalk.red(`SKIPPING extracting archives (due to config).`),
+			)
+			return
+		}
+		try {
+			await $`which ${extractcodeBinaryPath}`
+		} catch (e: unknown) {
+			this.logger.info(
+				chalk.red(`SKIPPING extracting archives (extractcode not found).`),
+			)
 			return
 		}
 
 		const verboseFlag = this.verbose ? '--verbose' : ''
-		const cmd = ['extractcode', verboseFlag, this.dirtyRoot]
+		const cmd = [extractcodeBinaryPath, verboseFlag, this.dirtyRoot]
 		await $`${cmd}`
 	}
 }
